+++ utility/Filter.h
@@ -64,7 +64,7 @@ inline GLint applyFilter(const filterType* filter, int filterW, int filterH, con
 			//convert coordinate to index in 1D
 			int filterPos = j * filterW + i; //According to blue book, pixels are stored row-first
 			int sourcePos = y * sourceW + x;
-			result += filter[filterPos] * source[sourcePos];
+			result += (int) (filter[filterPos] * source[sourcePos]);
 		}
 	}
 	return result;
